Refactor ort specific fusions
