+++ onnxscript/rewriter/__init__.py
@@ -6,7 +6,6 @@ from typing import Sequence, TypeVar, Union
 
 __all__ = [
     # Modules
-    "function_rule",
     "pattern",
     # Functions
     "rewrite",
@@ -16,18 +15,16 @@ import onnx
 
 from onnxscript import ir
 from onnxscript.optimizer import _remove_unused, _remove_unused_function
-from onnxscript.rewriter import function_rule, pattern
+from onnxscript.rewriter import pattern
 
 RewriteRuleSet = pattern.RewriteRuleSet
 PatternRewriteRule = pattern.RewriteRule
-FunctionRewriteRule = function_rule.FunctionRewriteRule
 
 ModelProtoOrIr = TypeVar("ModelProtoOrIr", onnx.ModelProto, ir.Model)
 
 
 def rewrite(
     model: ModelProtoOrIr,
-    function_rewrite_rules: Sequence[type[FunctionRewriteRule]] = (),
     pattern_rewrite_rules: Union[Sequence[PatternRewriteRule], RewriteRuleSet] = (),
 ) -> ModelProtoOrIr:
     if isinstance(model, onnx.ModelProto):
@@ -36,11 +33,6 @@ def rewrite(
     else:
         model_ir = model
         proto = False
-    if function_rewrite_rules:
-        for rule_cls in function_rewrite_rules:
-            count, model_ir = rule_cls().apply_to_model(model_ir)
-            if count > 0:
-                print(f"Applied {count} of rewrite rules.")
     if pattern_rewrite_rules:
         if not isinstance(pattern_rewrite_rules, RewriteRuleSet):
             # Create a pattern rule-set using provided rules
