+++ onnxscript/rewriter/onnxruntime/README.md
@@ -1,0 +1,1 @@
+This folder (and function_rule based rewrites) are deprecated. The folder will be removed soon.
