+++ onnxscript/rewriter/onnxruntime/transformers/multihead_attention.py
@@ -59,7 +59,7 @@ import onnxscript
 import onnxscript.ir.convenience
 import onnxscript.rewriter._ir_utils as _ir_utils
 from onnxscript import ir
-from onnxscript.rewriter import function_rule
+from onnxscript.rewriter.onnxruntime import function_rule
 
 logger = logging.getLogger(__name__)
