+++ onnxscript/rewriter/ort_fusions/fused_matmul_rule_sets_test.py
@@ -10,7 +10,7 @@ import onnx
 import onnx.reference
 import onnx.reference.op_run
 
-import onnxscript.rewriter.onnxruntime.fused_matmul_rule_sets as fused_matmul_rule_sets
+import onnxscript.rewriter.ort_fusions.fused_matmul_rule_sets as fused_matmul_rule_sets
 from onnxscript import ir
 
 FLOAT = onnx.TensorProto.FLOAT
